This is the current state of my code.

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/RobotInfo.java
@@ -1,7 +1,7 @@
 package org.firstinspires.ftc.teamcode;
 
 public class RobotInfo {
-    public float x;
-    public  float y;
-    public float degrees;
+    public double x;
+    public double y;
+    public double degrees;
 }
